+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Ensure `<AppIcon>` is not cut off when app name is long. Refs STCOR-752.
 * Allow console to be preserved on logout. STCOR-761.
 * Export `unregisterServiceWorker` to eliminate zombie service workers. Refs FOLIO-3627.
-* Fix duplicated "FOLIO" in document title in some cases. Refs STCOR-767.
+* Use keycloak URLs in place of users-bl for tenant-switch. Refs US1153537.
 
 ## [10.0.0](https://github.com/folio-org/stripes-core/tree/v10.0.0) (2023-10-11)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v9.0.0...v10.0.0)